Create Rollercoaster 2.0
This program will define if a user can or not ride based on his height. Considering the minimum height equals to 120 cm. In case the user is eligible to ride the rollercoaster, it will ask for his/her age and will print the corresponding price. 
In case the user wants to take a picture while on rollercoaster they must pay more $3, where the total bill will equal to age based price + photo price.